backend/requirements.txt
# fastapi
# uvicorn
# sqlalchemy
# bcrypt
# python-jose[cryptography]
# pydantic
fastapi
uvicorn
supabase
pydantic
python-dotenv
